refactor: extract useSettingsForm hook from settings shell
Move all form state, dirty tracking, and per-tab save handlers from
app/settings/index.tsx into a new lib/useSettingsForm hook. The shell
now owns only layout, deeplink (?tab=) parsing, tab switcher UI,
sticky save-bar wiring, and rendering of the active tab.

Pure structural refactor — no behavior changes, no UX changes, no
changes to the four tab components (Profile/Specialist/Notifications/
Account already extracted in PR #1496).

LOC: app/settings/index.tsx 591 -> 277 (under 500 soft-limit).
tsc --noEmit: 0 errors in both / and /api.
